Rename onDeviceMethod to onMethod for ModuleClient
