+++ device/core/test/_module_client_test.js
@@ -4,6 +4,7 @@
 'use strict';
 
 var fs = require('fs');
+var util = require('util');
 var EventEmitter = require('events').EventEmitter;
 var assert = require('chai').assert;
 var sinon = require('sinon');
@@ -11,6 +12,7 @@ var FakeTransport = require('./fake_transport.js');
 var Message = require('azure-iot-common').Message;
 var ModuleClient = require('../lib/module_client').ModuleClient;
 var errors = require('azure-iot-common').errors;
+var results = require('azure-iot-common').results;
 var ModuleClient = require('../lib/module_client').ModuleClient;
 var IotEdgeAuthenticationProvider = require('../lib/iotedge_authentication_provider').IotEdgeAuthenticationProvider;
 var SharedAccessKeyAuthenticationProvider = require('../lib/sak_authentication_provider').SharedAccessKeyAuthenticationProvider;
@@ -705,6 +707,197 @@ describe('ModuleClient', function () {
       client.setOptions(fakeOptions);
       assert.isTrue(methodClient.setOptions.calledOnce);
       assert.isTrue(methodClient.setOptions.calledWith(fakeOptions));
+    });
+  });
+
+
+  describe('#onMethod', function () {
+    var FakeMethodTransport = function (config) {
+      EventEmitter.call(this);
+      this.config = config;
+
+      this.connect = function (callback) {
+        callback(null, new results.Connected());
+      }
+
+      this.disconnect = function (callback) {
+        callback(null, new results.Disconnected());
+      }
+
+      // causes a mock method event to be raised
+      this.emitMethodCall = function (methodName) {
+        this.emit('method_' + methodName, {
+          methods: { methodName: methodName },
+          body: JSON.stringify(''),
+          requestId: '42'
+        });
+      };
+
+      this.onDeviceMethod = function (methodName, callback) {
+        this.on('method_' + methodName, callback);
+      };
+
+      this.sendMethodResponse = function (response, done) {
+        response = response;
+        if (!!done && typeof (done) === 'function') {
+          done(null);
+        }
+      };
+
+      this.enableMethods = function (callback) {
+        callback();
+      };
+
+      this.disableMethods = function (callback) {
+        callback();
+      };
+    };
+    util.inherits(FakeMethodTransport, EventEmitter);
+
+    // Tests_SRS_NODE_MODULE_CLIENT_13_020: [ onMethod shall throw a ReferenceError if methodName is falsy. ]
+    [undefined, null].forEach(function (methodName) {
+      it('throws ReferenceError when methodName is "' + methodName + '"', function () {
+        var transport = new FakeMethodTransport();
+        var client = new ModuleClient(transport);
+        assert.throws(function () {
+          client.onMethod(methodName, function () { });
+        }, ReferenceError);
+      });
+    });
+
+    // Tests_SRS_NODE_MODULE_CLIENT_13_024: [ onMethod shall throw a TypeError if methodName is not a string. ]
+    [new Date(), 42].forEach(function (methodName) {
+      it('throws TypeError when methodName is "' + methodName + '"', function () {
+        var transport = new FakeMethodTransport();
+        var client = new ModuleClient(transport);
+        assert.throws(function () {
+          client.onMethod(methodName, function () { });
+        }, TypeError);
+      });
+    });
+
+    // Tests_SRS_NODE_MODULE_CLIENT_13_022: [ onMethod shall throw a ReferenceError if callback is falsy. ]
+    [undefined, null].forEach(function (callback) {
+      it('throws ReferenceError when callback is "' + callback + '"', function () {
+        var transport = new FakeMethodTransport();
+        var client = new ModuleClient(transport);
+        assert.throws(function () {
+          client.onMethod('doSomeTests', callback);
+        }, ReferenceError);
+      });
+    });
+
+    // Tests_SRS_NODE_MODULE_CLIENT_13_025: [ onMethod shall throw a TypeError if callback is not a Function. ]
+    ['not_a_function', 42].forEach(function (callback) {
+      it('throws ReferenceError when callback is "' + callback + '"', function () {
+        var transport = new FakeMethodTransport();
+        var client = new ModuleClient(transport);
+        assert.throws(function () {
+          client.onMethod('doSomeTests', callback);
+        }, TypeError);
+      });
+    });
+
+    // Tests_SRS_NODE_MODULE_CLIENT_13_001: [ The onMethod method shall cause the callback function to be invoked when a cloud-to-device method invocation signal is received from the IoT Hub service. ]
+    it('calls callback when C2D method call arrives', function (done) {
+      // setup
+      var transport = new FakeMethodTransport();
+      var client = new ModuleClient(transport);
+      client.open(function () {
+        client.onMethod('firstMethod', function () { }); // This will connect the method receiver
+        client.onMethod('reboot', function () {
+          done();
+        });
+      });
+
+      // test
+      transport.emitMethodCall('reboot');
+    });
+
+    // Tests_SRS_NODE_MODULE_CLIENT_13_003: [ The client shall start listening for method calls from the service whenever there is a listener subscribed for a method callback. ]
+    it('registers callback on transport when a method event is subscribed to', function () {
+      // setup
+      var transport = new FakeMethodTransport();
+      var client = new ModuleClient(transport);
+      var callback = sinon.spy();
+      transport.on('newListener', callback);
+
+      // test
+      client.onMethod('reboot', function () { });
+
+      // assert
+      assert.isTrue(callback.withArgs('method_reboot').calledOnce);
+
+      // cleanup
+      transport.removeListener('newListener', callback);
+    });
+
+    // Tests_SRS_NODE_MODULE_CLIENT_13_023: [ onMethod shall throw an Error if a listener is already subscribed for a given method call. ]
+    it('throws if a listener is already subscribed for a method call', function () {
+      // setup
+      var transport = new FakeMethodTransport();
+      var client = new ModuleClient(transport);
+      client.onMethod('reboot', function () { });
+
+      // test
+      // assert
+      assert.throws(function () {
+        client.onMethod('reboot', function () { });
+      });
+    });
+
+    it('emits an error if the transport fails to enable the methods feature', function (testCallback) {
+      var transport = new FakeMethodTransport();
+      var fakeError = new Error('fake');
+      sinon.stub(transport, 'enableMethods').callsFake(function (callback) { callback(fakeError); });
+      var client = new ModuleClient(transport);
+      client.on('error', function (err) {
+        assert.strictEqual(err, fakeError);
+        testCallback();
+      })
+      client.onMethod('firstMethod', function () { }); // This will connect the method receiver
+    });
+  });
+
+  describe('transport.on(\'disconnect\') handler', function () {
+    var fakeTransport, fakeRetryPolicy;
+    beforeEach(function () {
+      fakeRetryPolicy = {
+        shouldRetry: function () { return true; },
+        nextRetryTimeout: function () { return 1; }
+      };
+
+      fakeTransport = new EventEmitter();
+      fakeTransport.enableMethods = sinon.stub().callsArg(0);
+      fakeTransport.onDeviceMethod = sinon.stub();
+    });
+
+    /*Tests_SRS_NODE_MODULE_CLIENT_16_099: [If the transport emits a `disconnect` event while the client is subscribed to direct methods the retry policy shall be used to reconnect and re-enable the feature using the transport `enableTwinDesiredPropertiesUpdates` method.]*/
+    it('reenables device methods after being disconnected if methods were enabled', function () {
+      var client = new ModuleClient(fakeTransport);
+      client.setRetryPolicy(fakeRetryPolicy);
+      client.onMethod('method', function () { });
+      assert.isTrue(fakeTransport.enableMethods.calledOnce);
+      fakeTransport.emit('disconnect', new errors.TimeoutError()); // timeouts can be retried
+      assert.isTrue(fakeTransport.enableMethods.calledTwice);
+    });
+
+    /*Tests_SRS_NODE_MODULE_CLIENT_16_100: [If the retry policy fails to reestablish the direct methods functionality a `disconnect` event shall be emitted with a `results.Disconnected` object.]*/
+    it('emits a disconnect event if reenabling methods fails', function (testCallback) {
+      var fakeError = new Error('fake');
+      var client = new ModuleClient(fakeTransport);
+      client.on('disconnect', function (err) {
+        assert.instanceOf(err, results.Disconnected);
+        assert.strictEqual(err.transportObj, fakeError);
+        testCallback();
+      });
+
+      client.setRetryPolicy(fakeRetryPolicy);
+      client._maxOperationTimeout = 1;
+      client.onMethod('method', function () { });
+      assert.isTrue(fakeTransport.enableMethods.calledOnce);
+      fakeTransport.enableMethods = sinon.stub().callsArgWith(0, fakeError);
+      fakeTransport.emit('disconnect', new errors.TimeoutError()); // timeouts can be retried
     });
   });
 });
