+++ device/samples/simple_sample_module.js
@@ -4,12 +4,12 @@
 'use strict';
 
 var Protocol = require('azure-iot-device-mqtt').Mqtt;
-var Client = require('azure-iot-device').Client;
+var ModuleClient = require('azure-iot-device').ModuleClient;
 var Message = require('azure-iot-device').Message;
 var fs = require('fs');
 var connectionString = process.env.EdgeHubConnectionString;
 
-var client = Client.fromConnectionString(connectionString, Protocol);
+var client = ModuleClient.fromConnectionString(connectionString, Protocol);
 console.log('got client');
 
 client.on('error', function (err) {
