+++ device/samples/simple_sample_module_twin.js
@@ -3,13 +3,13 @@
 
 'use strict';
 
-var Client = require('azure-iot-device').Client;
+var ModuleClient = require('azure-iot-device').ModuleClient;
 var Protocol = require('azure-iot-device-mqtt').Mqtt;
 var fs = require('fs');
 
 var connectionString = process.env.EdgeHubConnectionString;
 
-var client = Client.fromConnectionString(connectionString, Protocol);
+var client = ModuleClient.fromConnectionString(connectionString, Protocol);
 console.log('got client');
 
 client.on('error', function (err) {
